js/xpconnect/tests/mochitest/test_bug726125.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 726125</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank"
href="https://bugzilla.mozilla.org/show_bug.cgi?id=726125">Mozilla Bug 726125</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>

<pre id="test">
<script type="application/javascript">

SimpleTest.waitForExplicitFinish();

window.addEventListener("message", function(event) {
  /** Test for Bug 726125 **/
  ok(event.data == "Success", event.data);
  SimpleTest.finish();
}, false);

</script>
</pre>

<iframe src="http://www.mozilla.org/tests/js/xpconnect/tests/mochitest/helper_bug726125.html">
</iframe>

</body>
</html>
